backend/requirements.txt
# Aurevia Estate AI - Backend dependencies

fastapi==0.115.5
uvicorn[standard]==0.32.1

pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0

sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg[binary]==3.2.3
alembic==1.14.0

openai==1.57.2
qdrant-client==1.12.1

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

httpx==0.28.1
aioimaplib==2.0.1

python-dotenv==1.0.1
python-multipart==0.0.20
tenacity==9.0.0

pytest==8.3.4
pytest-asyncio==0.24.0
